pyproject.toml
[project]
name = "csiro-image2biomassprediction"
version = "0.1.0"
description = ""
authors = [
    { name = "Ivan Bashtovyi", email = "ivanbashtovyi1@gmail.com" },
    { name = "Arthur Khodakovskyi", email = "arturkhodalovskiy@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = ["kaggle (>=1.7.4.5,<2.0.0.0)", "torch (>=2.9.0,<3.0.0)", "numpy (>=2.3.4,<3.0.0)", "pandas (>=2.3.3,<3.0.0)", "scikit-learn (>=1.7.2,<2.0.0)", "transformers (>=4.57.1,<5.0.0)", "hydra-core (>=1.3.2,<2.0.0)", "pillow (>=12.0.0,<13.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"